微信业务域名回调txt

+++ atjx/src/main/java/com/atjx/model/Item.java
@@ -2,7 +2,6 @@ package com.atjx.model;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.Date;
 
 @Data
@@ -14,7 +13,7 @@ public class Item extends BaseObject {
     //商品详情
     private String sellPoint;
     //商品价格
-    private BigDecimal price;
+    private Integer price;
     private int num;
     private String barcode;
     private String image;
@@ -25,13 +24,13 @@ public class Item extends BaseObject {
     private String createdStr;
     private String updatedStr;
     private String categoryName;
-    private BigDecimal minPrice;
-    private BigDecimal maxPrice;
+    private Integer minPrice;
+    private Integer maxPrice;
     private Integer minNum;
     private Integer maxNum;
-    private BigDecimal del_price;
+    private Integer del_price;
     private int sales;
-    private BigDecimal money;
+    private Integer money;
     private Address address;
     private Specification specification;
